# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Planteo",
        "description": "API pour gérer les utilisateurs, les plantes, et les plantes favorites",
        "version": "1.0.0"
    },
    "paths": {
        "/api/favorites": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get the favorites of the connected user",
                "operationId": "d7a4e876a8d5f9815295c7fc4954f011",
                "responses": {
                    "200": {
                        "description": "List of favorite plants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plant"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a plant to favorites",
                "operationId": "370ffab0458b21c048b584dc20d059e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plant added to favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante ajoutée à vos favoris."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Plant already in favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cette plante fait déjà partie de vos favoris"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove a plant from favorites",
                "operationId": "fa0c59fb9c161b977258ec7656104ab2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plant removed from favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante supprimée de vos favoris."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found in favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La plante n'est pas dans vos favoris"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorite/{id}": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Check if a plant is in favorites",
                "operationId": "b288def8554521e1b1d407a4b1c4ddb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isFavorite": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Plant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Haricots verts"
                    },
                    "description": {
                        "type": "string",
                        "example": "Les haricots verts sont des légumes savoureux et croquants, parfaits pour des salades ou des plats cuisinés."
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "type": {
                        "type": "string",
                        "example": "Légume vert"
                    },
                    "sowing_period": {
                        "type": "string",
                        "example": "03, 04"
                    },
                    "planting_period": {
                        "type": "string",
                        "example": "05, 06"
                    },
                    "harvest_period": {
                        "type": "string",
                        "example": "07, 08, 09"
                    },
                    "soil": {
                        "type": "string",
                        "example": "Sol léger"
                    },
                    "watering": {
                        "type": "string",
                        "example": "Arrosage important pendant la floraison"
                    },
                    "exposure": {
                        "type": "string",
                        "example": "Plein soleil"
                    },
                    "maintenance": {
                        "type": "string",
                        "example": "Biner régulièrement, pailler pour maintenir l'humidité, et butter les plants lorsqu'ils atteignent 10-15 cm"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}